Fix LaTeX images everywhere
I realize it's probably best to avoid adding things to the "*" section, but I keep enabling this rule locally using the dev tools and I'm not sure what else to do. I don't want to spam PRs or fill up `inversion-fixes.config` with this same rule for a dozen different sites. This appears to be a very common way to embed LaTeX on a page and I can't imagine inverting `img.latex` is going to break very much (how many non-LaTeX images have a `.latex` class?).