Old Reddit title color
Reddit redesigned their site to look different a while ago but the old design is still an option under a user's preferences.  This inverts the color for post titles when a user has the old design set in their preferences.  It has no effect on the new design.  Without this, post titles are hard to read since they are a very dark blue on a dark background.