Fix inversion for Instagram and FiveThirtyEigth
For Instagram, I fixed inversion of logo and the top panel so that they
are visible. The website looks better since now we don't have black
title on gray background.
I made similar change to FiveThirtyEight. The title of the website
should be visible as well as the searchbar. This change makes them
visible.