pretty good scrollbar
The following is from google translation:
I am looking for a scrollbar that works better in dark conditions. At first I used tampermonkey, work well, but I was confused about the static and dynamic processing of the dark reader. Then I found that the developer tool can modify the dynamic theme, and directly get the scrollbar into it.