Fix browser theme setting in Firefox
The new `frame` and `tab_background_text` properties must be set in
order to avoid setting to defaults in newer Firefox versions (e.g. on
Nightly).

At the moment the old `accentcolor` and `textcolor` properties return a
warning, so we can avoid removing them to work with older versions of
Firefox for as long as possible.

Fixes: https://github.com/darkreader/darkreader/issues/1430
Refs: https://github.com/darkreader/darkreader/commit/181ec067486ab4602d6d1c78ed5be9eac089139d