Update DTF for YNAB Calendar Icon (#1604)
On YNAB, if you set up a recurring transaction for a category, an SVG calendar icon shows up next to the available amount for that category on the Budget screen. Dark Reader by default turns the color of the icon to the same as the surrounding background color, making it essentially invisible. This change fixes the issue and keeps the color of the calendar icon white.