Fix category refinement for JCrew
When refining categories on JCrew's website, all checkboxes were orange, regardless of whether or not they were selected. The only difference was that selected items had a black background, and nonselected ones had a gray background. It's nearly impossible to tell the difference.

Now, the checkboxes are invisible if they are not selected, and white if they are selected. The black vs. gray background remains untouched.