Partial fix for Microsoft main logo
[microsoft.com](https://microsoft/com)
Inverting the logo kind of works, but the background is black, not gray like the rest of the page (and it is a PNG, so can't fix background with CSS).

Also, when using light mode the image is now broken again, but the opposite of what it was before. I don't know how to fix this, it would be nice to just remove the background.

Also also, it seems like people have attempted to fix the Microsoft logo at least three times before, but they keep updating the image version. This is more of a temporary fix, I might see if I can do a more complex selector later.