Added a fix for heise.de
I added a fix for the german technology-news-site "heise.de": It wasn't possible to see the results of the opinion-section of some news (a section where readers can vote for their opinion) because the frame was incorrectly inverted. I changed this behavior just by adding an "INVERT" rule, tested it in both modes and this small tweak works really good.