Prevent whiteflash on inital document_start - Adds a whiteflash killer style on document_start instead of checking all variables and do go trough functions. - Adds fallback style a lot earlier to preven white flash.
Even on a high-end system I noticed a nice white flash, however with this I didn't notice it. Also chrome has a nice feature to reduce memory usage and free memory of unused sites and give a even longer whiteflash, not sure how to fix that or if chrome need to do some magic to make sure to get it's dark-ish color.