Correct some image elements on artofproblemsolving.com
The site uses images mixed in with text for some special characters. The images have transparent backgrounds and black symbols on them, and hence they become black on transparent background when inverted. When rendered on the webpage, they are nearly the same color as the background, and are hence unreadable This change fixes those inversions.
The example page I visited: https://artofproblemsolving.com/wiki/index.php/2010_AMC_10A_Problems