Update dark-sites.config
My website "benfoster.dev", along with the in-progress version already bares a full dark design without the ability to modify to a light version dynamically. When dark reader is enabled, the extension improperly modifies the logo constructed via HTML and CSS. Observe the brackets on the left and right becoming black, along with the backdrop-filter blur properties being disabled.