Added a fix for Aftonbladet.se
Aftonbladet.se is one of the largest newspapers in Sweden. I added a fix for a button on the main page, which can be seen on all sub-pages. There was a contrast error, that had to do with Darkreader inverting black text to white. 

Will probably find more things to correct as I use the website more in the future.