Fix code diff and terminal font on gitlab.com
As of the important flag added to the `:not(pre)` selector and because GitLab.com doesn't use
pre/code elements for individual code and terminal lines, the font family in terminals and code diffs
are overwritten in case the font settings in Dark Reader are changed, e.g. to the system font.
Re-applying GitLab's font-stack to those elements again, with the !important flag, solves the issue.