Fix for dark themes
This is a fix that needs to be carefully considered before accepting, as it affects global elements.

I have found this fix required when using dark themes in the Windows OS and using Firefox.  Without this fix, HTML input, select, and textarea fields will sometimes be impossible to read, as they will be white on white.

I have been running with this fix for several months without any issues (and with it working successfully to resolve the aforementioned issue), but I have not performed any testing outside of my environments.