Fix dark filter (inversion) mode
This is another fix that needs to be carefully considered before accepting, as it affects global elements.

I have found this fix required when using dark themes in the Windows OS and using Firefox.  Without this fix, HTML input, select, and textarea fields will sometimes be impossible to read, as they will be white on white.