Color inversions for 'move' menu elements
When clicking the move icon (between the star and cloud icons) at the top, a small menu window ui lets you choose the destination. Some of the following icon elements were not inverted and I'm adding them to INVERT:
- Open folder in new window
- Window closing "x"
- Create new folder
- Select folder right chevron (hover)
- Shared with me (at root)
- Starred (at root)

Also added to CSS a bright color for the scrollbar thumb for this little window.