Properly show content on whiteboard
Invert content on the interactive whiteboard on suite.smarttech-prod.com. Before the change, the white background was changed to dark, but the content (such as text) was still black, making it hard to read. This was the only way I found to fix it, by just inverting the content.

Before change:
[https://imgur.com/a/dBV7uU1](url)

After change:
[https://imgur.com/a/txB0dep](url)