Add fixes for Google Classroom
On any Google Classroom, in the "Classwork" tab, there is a list, and each item of the list is some material or assignment. A list item gets highlighted when is it clicked or hovered on.
When an item is clicked (and thus highlighted) and the item above it hovered on, they both normally get the same color; but with Dark Reader, the lower one goes completely blank.

With some experimentation, I came to the conclusion that giving the two highlighted nodes different colors solves the problem; I still don't know why though.
This change leaves the opacity of the hovered-over list item at 100%; all the remaining list items are given an opacity of 99%. This creates the required color difference without drastically changing the look to the user.