Use Alarms API for reconnecting socket on dev build (#6327)
This commit replaces one more instance of setTimeout in the background context with an alarm. This is direct continuation of work done in #6296.

Note: This alarm runs in 1s (like the original code) and browsers can log warnings promising to throttle alarms in release builds to once a minute. This is not a problem, since this code will never run in release builds.