Use crypto.randomUUID() in Utils UID (#6387)
Chrome 92 added support for crypto.randomUUID() function which returns formatted hex-encoded Universally Unique Identifiers (UUID) based on RFC4122. The new generateUID implementation based on crypto.randomUUID() is approximatelly 2x faster than the old one based on crypto.getRandomValues().

Note: this method leads to reduced entropy of the resulting string (by 6 bits) because crypto.randomUUID() produces UUID with predefined version identifiers. This entropy reduction is not critical given the overall entropy of the string and the way this function is used.