Created and proposed Filipino translation
According to the [latest Statista report](https://www.statista.com/statistics/1247796/internet-browser-market-share-philippines/), 81% of the internet users in the Philippines use Google Chrome, one of the browsers Dark Reader supports. By localizing the extension to Filipino, we can attract more users and contributors from the Philippines. This may result in Dark Reader gaining a better reputation across Southeast Asia. I have checked that my translations are as accurate and as short as possible, and I have put lots of effort into it. I hope that my translations will be accepted. 😊