Make Toucan's pinyin field visible
When using the Toucan add-on for language learning, at least for Chinese, selecting a translated word provides both the translation and the pronunciation in pinyin. Pinyin section is almost unreadable.

This fix changes the text color to neutral, so pinyin is readable again. Probably works for other languages, but I only tested Chinese.