Improved conversions on colors.dopely.top
I made some changes for this site where the shades of colors are displayed correctly, as well made the tool icon symbols more visible.

NOTE: There is a bug that if you load a page on site with darkreader already on, things are displayed incorrectly for some reason. However, turning off and on darkreader instantly fixes it.