Partial Fix for wacom.com
This fixes the logos at the top and bottom of the page. This site does a lot of the pictures with words on top of it thing. Dark Reader flips the text to white, so they are white text on top of white backgrounded images. I'm not sure what to do about this, and I don't want to just make them dark text on white background, so I've done nothing about that part, even though as is, the text is not readable mostly everywhere.