Fix Slack.com dynamic theme
There are two issues with Slack.com's current dynamic fixes:
  - The invert logo on the homepage no longer works (the logo is now an SVG, and has a newly named class)
  - Buttons in the actual "app" aren't darkened properly

The first issue is resolved by coloring the first path of the logo's SVG element. This is the only way to select the text part of the logo as they have not named the paths, and the colorful "pips" of the logo are part of the same SVG element.

The second issue is resolved by letting the buttons inherit their text color and background. This has the unfortunate side effect that some buttons which were previously "colored" (such as Like buttons being green) are now greyer, but it was the only solution I could find that works in both light and dark schemes. With that change, there is no indication when hovering, so I added a basic underline decoration for that. I also took care not to enable it for primary and secondary buttons which seem to be handled fine by the dynamic mode.