Add a Immediate modify option
- Add a per-site option to enable "Immediate modify", within this
mode(only useable for dynamic theme), Dark Reader will modify the page
without waiting for it to be shown to the user. With this mode you can
open a new tab in the background and switch to it and notice it's
already been modified by Dark Reader.
- Because users(myself included to a certain extend) might find this
behavior unexpected and prefer to let Dark Reader wait until the tab is
shown, this option is disabled by default. And because it has chances it
will break certain sites, it's implemented to be a per-site toggle(Under
`font & more`).
- Resolves #7682.