Use Shunting-Yard algorithm for lowering calc.
- Currently we were using a hacky made up parser(made by me) to handle
simple cases of `calc(...)` and safely evaluate them. However this
doesn't handle all cases and updating this hacky function to also handle
precende is not great for maintaince.
- This patch fixes that by using the [Shunting-Yard
algorithm](https://en.wikipedia.org/wiki/Shunting_yard_algorithm). Which
is a small but powerful parser & evaluator which converts infix
notation and transform it into RPN(which is really really easy and fast to evaluate).
- Resolves #8894