Only send messages to affected tabs when toggling current site
- Currently if you toggle a site, a new tabmessage would be sent to all
tabs, this is inefficient as most sites will just reprocess the whole
site without any changes in the output. This PR fixes that by currently
hardcoding to only send tabmessages to tabs that have the same hostname
as the active tab, when toggling the current site.
- This behavior can later extended to other functionality, however this
requires processing of presets, themes etc. Which require the more
complex URL matching.
- Resolves #9336