Restore functional colors on Manifold Markets
Many of the colors necessary for interpreting information on Manifold are missing when using Dark Reader's "Dynamic" mode. Most importantly, Manifold's green buttons fail to function properly, especially when :hover selectors are involved. This PR restores the functionality of those buttons, as well as those involved with canceling limit orders.