Add background handling for lectricebikes.com
I'm not very familiar with HTML, but it appears there are some variables in the page source e.g. `--gradient-base-background-1: #FFFFFF;` and changing it to darkreader backgrounds fixes the background. It's within `<head> <style data-shopify="">` I can't figure out how to make the text other than the multiple shades of blue, my attempt is in there (original values for --gradient-base-accent match the darkreader values for --blue and --ltblue).

My first ever commit, so please bear with me.