Update dynamic-theme-fixes.config
Made the "level bars" visible on dark mode. 

There are a handful of w3.org SVGs that I didn't know the solution for however. A couple until you scroll down to the "effects" section, where there are 9 grouped up together. If someone does have the solution for this, I would be interested in learning it!