Update dynamic-theme-fixes.config
Added support for rewards.microsoft.com and rewards.bing.com. Right now rewards.microsoft.com redirects to rewards.bing.com, but they are the same site, Microsoft just changed the URL to bing in the last day or so.
The IGNORE INLINE STYLE is there to make sure the various svg icons and animations they have on the page stay their original color, since they can look weird when modified.
The INVERT is to make some text remain dark in so that it can be read in front of a bright image behind it.