Update dynamic-theme-fixes-config for bol.de
On bol.de pictures aren't shown for search results.
Example: https://www.bol.de/kategorie/brettspiele-4810/

I changed the z-index of the shadow-box-wrapper so the picture is in foreground. This disables the option to click the picture, but at least it shows.
Unfortunately i'm not well-versed in CSS, so this is the best i can do.