Fix for kaava.mafynetti.com
Fix for a math input site used in almost every finnish high school. 

Fixed so that you can see the text on the math input images, as it was black on a really dark background before which made this extension unusable on this website.
It was the first time I tried changing anything on a website, and I didn't manage to change the color of the text in the image, so I managed to change the background in the images instead. 

Someone else could probably do this a lot better than me, but I don't think that is going to happen and this is much better than completely unusable :D