pixilart.com gallery images fix and logo fix
Reverting image inversion of gallery images. In some cases one can hide their "light mode" logo and replace it with their in-built "dark-mode" one. In other cases (the homepage) there does not exist a second element to swap to. Instead I change the src to point to the "dark mode" image they host.