Fix claude.ai top bar and logo button (#11506)
Fixed small glitches of the dynamic theme on https://claude.ai/

When in conversation with AI, the top white bar isn't inverted but the text on it is inverted to white, so the text is not so readable. I adjusted the CSS but there is still some half-transparent glow effect. I don't know where it comes, but at least the text is readable after fixing.

Also when in a conversation with the AI, the top-left logo button is not inverted correctly. The button background color is inverted but the image in the button isn't inverted.