Update dynamic-theme-fixes.config for Web of Science
Web of Science search function dynamically marks search text with yellow background color.
This creates bad contrast with default color of article headings in purple text color.
Which makes reading the highlighted text difficult to read in dark mode.

This fix is to solve the visibility problem. 
Specifically changes the color of text that is dynamically marked to darkreader-neutral-text.