Add circuits.vcn.com to dynamic-theme-fixes.config
This makes images on a site that my company uses for work more readable in dark mode. It also makes selection boxes readable. Since this is a private access website, I understand if this won't be accepted, but I'm not sure how to make the changes automatically apply when I open up the site for the first time without having to go into dev tools and apply them there. Other users of the site at my company would appreciate this very much, since not everyone is technical enough to figure this out. Let me know if you want pictures of how this looks in dark mode and light mode, or if you need anything else.