Fix LaTeX code color in pir2.forumeiros.com and app.estuda.com (#12981)
* Fix LaTex code color in pir2forumeiros.com and app.estuda.com

The LaTeX images are being displayed in black color, and considering the dark background, it's very difficult to understand them. The proposed change aims to invert the color of these images to white, as to make them easy to read.

Examples of the change on these sites can be seen in:

https://pir2.forumeiros.com/t205435-probabilidade-banca-fgv#705383

https://app.estuda.com/questoes/?id=11473523

* Update dynamic-theme-fixes.config

Fix alphabetical order of fixes.