Fix spaces in code blocks on palletsprojects.com (#14409)
* Fix spaces in code blocks on palletsprojects.com

Some of the spaces in code blocks on *.palletsprojects.com appear as white underscores in dark mode, which makes the code more difficult to read. You can see this in the first code block on https://flask.palletsprojects.com/en/stable/async-await/. There should be a space, not an underscore, between `def` and `get`.

* Fix alphabetization