Add dark site and adjust README (#14529)
This PR adds a dark site to the list and adjusts the README to use a permanent link for the Firefox's source code (otherwise, the link position gets invalidated after new commits).

The site added to the list is a quest mini-game that is played via the browser, and Dark Reader makes some parts of it invisible.