Fix for www.PlayInterference.com (#14540)
* Fix for www.PlayInterference.com

This is for www. and not *. as legacy. exists

This a drawing game - the colour palette gets modified by DarkReader, and on phone interface some dots appear with also same wrong colours for "border" on colour picker, so what you see isn't what you're drawing. This CSS fixes that and the same colours applies to .p-1 brushes preview. However, to avoid doubling the CSS, the "border-width: 0px" for colour picker squares also applies to .p-1 brushes' DIVs, which makes the brushes borders invisible (already 1px), and adding a new ".p-1 div{border-width: 1px !important}" at the end doesn't override that. However, the brush preview should contrast enough with the dark bg which has been made sligthly clear (it was really dark and merged with the real BG and black brush preview).

* Fix for www.PlayInterference.com

Improvement over the previous commit. This is now fully fixed.

* Fix for www.PlayInterference.com

Moved URL from "pla..." to "www.pla..."

* Update dynamic-theme-fixes.config

Moved back to "pla..." (from "www.pla...") and removed "www."

* Update dynamic-theme-fixes.config

Forced real black for selected-colour instead of "black" which DarkReader turned into grey on real phone but not when emulating on desktop.

* Update dynamic-theme-fixes.config

Fixed the black border turning grey - it got only black when pressing "apply" on the dev window, but not when loading the page with no user intervention, so I was getting tricked by that.