[http://swagger.emby.media] (emby media server api swagger UI), invert svg, banner color
the banner was the same color as the background.
the SVG down arrow to uncollapse were black, same as the locks and close modal x.
the `.view-line-link` is set on a button to copy to clipboard, inverted looks better IMO.

the CSS is for the down arrow of a select box.
i had to copy paste the SVG used by the original CSS and add the `fill="white"`. it now appears bigger, and I don't know why, nor do I care enough to know why. originally it was set to `background-image: url("blob:http://swagger.emby.media/eb40c362-ec66-409d-88a4-958bf5380c22");`, which means I can't modify it, or give some selector for an element, since it isn't one.

NOTE: to actually get something shown to which these rules apply, you'd need to enter the (probably locally hosted) URL of an [emby](https://emby.media/) server.