Update dynamic-theme-fixes.config (#14592)
* Update dynamic-theme-fixes.config

Updated the configuration to include the CSS selector targeting the bright master images on microsoft.com within the invert section of the microsoft.com entry, excluding the blue master image to prevent inversion from enhancing its brightness.

* Update dynamic-theme-fixes.config

The line was removed to prevent the occurrence of consecutive empty lines, which cause test failures.