Invert and fix pdfViewer in hypothes.is
Fixes need to apply to multiiple hosts, so switched to wildcard format. I suppose not everyone wants the PDF Viewer inverted, but for me it's a severe accessibility issue. I also had to fix some really bad CSS issues (yellow on white with an opacity of 0.4 and alpha of 0.3 = TOTALLY INVISIBLE to me, even inverted). If it is possible to apply these only if the user has opted to invert PDF's, that may be desirable.